+++ source_gen/source_gen/lib/source_gen.dart
@@ -7,3 +7,4 @@ library source_gen;
 export 'src/builder.dart';
 export 'src/generator.dart';
 export 'src/generator_for_annotation.dart';
+export 'src/type_checker.dart' show TypeChecker;